+++ tests/unit/test_watch_dog.py
@@ -1,6 +1,6 @@
 import threading
 import time
-from mserv.utils import watch_dog
+from socketlib.utils import watch_dog
 
 
 def function_that_raises_error():
@@ -10,7 +10,7 @@ def function_that_raises_error():
 
 
 def test_calls_exit_when_a_thread_dies(mocker):
-    mock_exit = mocker.patch("mserv.utils.watch_dog.os._exit")
+    mock_exit = mocker.patch("socketlib.utils.watch_dog.os._exit")
     thread = threading.Thread(target=function_that_raises_error, daemon=True)
 
     dog = watch_dog.WatchDog()